src/main/java/be/bonamis/advent/year2022/Day07.java
package be.bonamis.advent.year2022;

import be.bonamis.advent.DaySolver;
import be.bonamis.advent.utils.FileHelper;
import lombok.Getter;
import lombok.ToString;

import java.util.Arrays;
import java.util.Comparator;
import java.util.List;
import java.util.stream.Stream;

import static java.util.stream.Collectors.toList;

public class Day07 extends DaySolver<String> {

    public Day07(List<String> puzzle) {
        super(puzzle);
    }

    @Override
    public long solvePart01() {
        return this.puzzle.size();
    }

    @Override
    public long solvePart02() {
        return this.puzzle.size() + 1;
    }
}
